Added .txt files to .gitignore. Fixed some parsing issues
I initially assumed that all the items follow a key=value pair format.

The error came here : For example, if keyValuePair is "name=John=Doe", the code will split it based on the first = and assign "name" as the key and "John" as the value. The remaining = and "Doe" will be lost.

I modified the code to split the key-value pair by the first occurrence of =, instead of using split("=").